requirements.txt
pydantic~=1.9.0
aioify~=0.4.0
py-cpuinfo~=8.0.0
psutil~=5.9.0
motor~=2.5.1
PyYAML~=6.0
cachetools~=5.0.0